migrate logic into BaseTransform
